barrelOS: start work on 64-bit port

+++ kernel/lib/bitmap.h
@@ -3,15 +3,15 @@
 
 #include <stdint.h>
 
-void bitmap_set(uint8_t *bitmap, uint32_t bit) {
+void bitmap_set(uint8_t *bitmap, uint64_t bit) {
     bitmap[bit / 8] |= 1 << (bit % 8);
 }
 
-void bitmap_clear(uint8_t *bitmap, uint32_t bit) {
+void bitmap_clear(uint8_t *bitmap, uint64_t bit) {
     bitmap[bit / 8] &= ~(1 << (bit % 8));
 }
 
-bool bitmap_get(uint8_t *bitmap, uint32_t bit) {
+bool bitmap_get(uint8_t *bitmap, uint64_t bit) {
     return bitmap[bit / 8] & (1 << (bit % 8));
 }
